pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "faster-whisper>=1.2.1",
    "google-genai>=1.69.0",
    "moviepy>=2.2.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "sentence-transformers>=5.3.0",
    "uvicorn>=0.42.0",
]
